+++ lib/cocoapods/generator/embed_frameworks_script.rb
@@ -181,7 +181,7 @@ module Pod
           strip_invalid_archs() {
             binary="$1"
             # Get architectures for current target binary
-            binary_archs="$(lipo -info "$binary" | rev | cut -d ':' -f1 | awk '{$1=$1;print}' | rev)"
+            binary_archs="$(lipo -archs "$binary")"
             # Intersect them with the architectures we are building for
             intersected_archs="$(echo ${ARCHS[@]} ${binary_archs[@]} | tr ' ' '\\n' | sort | uniq -d)"
             # If there are no archs supported by this binary then warn the user