+++ CHANGELOG.md
@@ -32,6 +32,10 @@ To install release candidates run `[sudo] gem install cocoapods --pre`
 * Add support for integrating dependency file in user script phases.  
   [Dimitris Koutsogiorgas](https://github.com/dnkoutso)
   [#9082](https://github.com/CocoaPods/CocoaPods/issues/9082)
+ 
+* Add support for XCFrameworks using the `vendored_frameworks` Podspec DSL.  
+  [Eric Amorde](https://github.com/amorde)
+  [#9148](https://github.com/CocoaPods/CocoaPods/issues/9148)
 
 ##### Bug Fixes